fix: Resolve all CI pipeline errors and remove hardcoded API keys
- Fix 10 ruff lint errors (StrEnum, UP038, unused imports, import sorting)
- Fix 45 mypy type errors across 14 source files
- Run black formatting on all affected files
- Replace broken pkcs11 package with maintained python-pkcs11 fork
- Add quality-gate job to release workflow to prevent publishing with broken CI
- Move EXA_API_KEY from source files to .env (read via os.environ.get)
- Clean up .gitignore (deduplicate, add IDE/OS patterns)

+++ src/qkdpy/protocols/enhanced_cv_qkd.py
@@ -320,9 +320,11 @@ class EnhancedCVQKD(BaseProtocol):
         )
 
         avg_corr = (actual_corr_x + actual_corr_p) / 2
-        noise_estimate = max(0.0, (expected_corr - avg_corr) / expected_corr)
+        noise_estimate = float(
+            max(0.0, (expected_corr - avg_corr) / expected_corr)  # type: ignore[arg-type]
+        )
 
-        return float(noise_estimate)
+        return noise_estimate
 
     def _get_security_threshold(self) -> float:
         """Get the security threshold for the CV-QKD protocol.
